nfc-64 part 1: store native structure in a long rather than an int

+++ nci/src/com/android/nfc/dhimpl/NativeNfcManager.java
@@ -61,7 +61,7 @@ public class NativeNfcManager implements DeviceHost {
     public static final String INTERNAL_TARGET_DESELECTED_ACTION = "com.android.nfc.action.INTERNAL_TARGET_DESELECTED";
 
     /* Native structure */
-    private int mNative;
+    private long mNative;
 
     private final DeviceHostListener mListener;
     private final Context mContext;
